fix: access-client authorize and waitForDelegation use cases (#563)
Motivation:
* avoid adding new public method to Agent that can be a function that
operates on the existing public api
